# api/openapi.yaml
openapi: 3.1.0
info:
  title: SYNC Server API
  version: 0.1.0
  description: |
    Federated, privacy-focused 1:1 chat server.
    Access tokens are short-lived HS256 JWTs (1 h TTL).
    Refresh tokens are 30-day opaque tokens subject to rotation and family-replay detection.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TokenResponse:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
          description: HS256 JWT valid for `expires_in` seconds
        refresh_token:
          type: string
          description: Opaque 32-byte random token (hex-encoded, SHA-256 hashed in DB)
        expires_in:
          type: integer
          description: Seconds until the access token expires (3600)
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
paths:
...
  /auth/refresh:
    post:
      operationId: authRefresh
      summary: Rotate refresh token and get a new access token
      security: []
      tags: [auth]
      description: |
        Accepts a valid, unused refresh token and returns a new `TokenResponse`.
        The old token is immediately invalidated (rotation).
        If a previously-used token is presented, the entire token family is revoked
        (replay attack detection) and 401 is returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Tokens rotated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Token invalid, expired, revoked, or replayed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...